processorexample/db-analyzer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.32.5
beautifulsoup4==4.12.3
python-dotenv==1.0.0
openai==1.54.0
httpx==0.25.0
psycopg2-binary==2.9.9